Re-implement fork/exec logic in Daemon.daemonize in order to use the ProcessBuilder APIs
There is a case in the JNA call to fork where the forked process needs to wait on a safepoint in the JVM
after the native call. Because there are no other JVM threads to unlock this safepoint in the child process
operating on the separate memory, the thread will be locked forever, unable to make the execv call.

The fork and exec need to happen in the same native call in order for this to be safe.
Brief description of the problem for context: https://stackoverflow.com/a/51917935

It appears as though a fork (that isn't immediately followed by an execv) is not safe in any case on the
JVM.